[TMail] refactor: group all AiBotReplier related classes in one folder
