+++ tmail-backend/tmail-third-party/ai-bot/AIBotMailet.adoc
@@ -10,14 +10,14 @@ The `mailetcontainer.xml` file should be configured as follows:
 ----
 <matcher name="aibot-allowed" match="org.apache.james.mailetcontainer.impl.matchers.And">
       <matcher match="SenderIsLocal"/>
-      <matcher match="com.linagora.tmail.mailet.RecipientsContain=gpt@example.com"/>
+      <matcher match="com.linagora.tmail.aiBotReplier.RecipientsContain=gpt@example.com"/>
 </matcher>
 ...
 
 <mailet match="All" class="com.linagora.tmail.mailets.TmailLocalDelivery">
       <consume>false</consume>
 </mailet>
-<mailet match="aibot-allowed" class="com.linagora.tmail.mailet.AIBotMailet">
+<mailet match="aibot-allowed" class="com.linagora.tmail.aiBotReplier.AIBotMailet">
       <botAddress>gpt@example.com</botAddress>
 </mailet>
 <mailet match="All" class="Null"/>
