+++ tmail-backend/tmail-third-party/ai-bot/sample_conf/mailetcontainer.xml
@@ -89,7 +89,7 @@
         <processor state="local-delivery" enableJmx="true">
             <matcher name="aibot-allowed" match="org.apache.james.mailetcontainer.impl.matchers.And">
                 <matcher match="SenderIsLocal"/>
-                <matcher match="com.linagora.tmail.mailet.RecipientsContain=gpt@example.com"/>
+                <matcher match="com.linagora.tmail.aibotreplier.RecipientsContain=gpt@example.com"/>
             </matcher>
             <mailet match="All" class="VacationMailet"/>
             <mailet match="All" class="Sieve"/>
@@ -100,7 +100,7 @@
                 <consume>false</consume>
             </mailet>
 
-            <mailet match="aibot-allowed" class="com.linagora.tmail.mailet.AIBotMailet">
+            <mailet match="aibot-allowed" class="com.linagora.tmail.aibotreplier.AIBotMailet">
                 <botAddress>gpt@example.com</botAddress>
             </mailet>
             <mailet match="All" class="Null"/>
