+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/aibotreplier/RecipientsContain.java
@@ -16,7 +16,7 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.mailet;
+package com.linagora.tmail.aibotreplier;
 
 import java.util.Collection;
 import java.util.Optional;
